Move all test to 6.0 and delete 6.1 tests
