+++ dev/io.openliberty.webcontainer.servlet.6.1.internal_fat/.classpath
@@ -12,7 +12,6 @@
     <classpathentry kind="src" path="test-applications/HttpSessionAccessor.war/src"/>
     <classpathentry kind="src" path="test-applications/PathCanonicalizationTest.war/src"/>
     <classpathentry kind="src" path="test-applications/ReadWriteByteBuffer.war/src"/>
-    <classpathentry kind="src" path="test-applications/RequestCookieHeader61Test.war/src"/>
     <classpathentry kind="src" path="test-applications/RequestParameterTest.war/src"/>
     <classpathentry kind="src" path="test-applications/ResponseNoOpAfterCommit.war/src"/>
     <classpathentry kind="src" path="test-applications/ResponseSendRedirectTest.war/src"/>
