+++ dev/io.openliberty.webcontainer.servlet.6.1.internal_fat/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2023, 2026 IBM Corporation and others.
+# Copyright (c) 2023, 2024 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -23,7 +23,6 @@ src: \
     test-applications/HttpSessionAccessor.war/src,\
     test-applications/PathCanonicalizationTest.war/src,\
     test-applications/ReadWriteByteBuffer.war/src,\
-    test-applications/RequestCookieHeader61Test.war/src,\
     test-applications/RequestParameterTest.war/src,\
     test-applications/ResponseNoOpAfterCommit.war/src,\
     test-applications/ResponseSendRedirectTest.war/src,\
