--- dev/io.openliberty.webcontainer.servlet.6.1.internal_fat/publish/servers/servlet61_RequestCookieHeaderTest/bootstrap.properties
@@ -1,15 +1,0 @@
-###############################################################################
-# Copyright (c) 2026 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-###############################################################################
-bootstrap.include=../testports.properties
-osgi.console=7777
-com.ibm.ws.logging.trace.specification=*=info=enabled:io.openliberty.webcontainer*=all:com.ibm.ws.webcontainer*=all:com.ibm.wsspi.webcontainer*=all:HTTPChannel=all:GenericBNF=all:HTTPDispatcher=all:com.ibm.ws.http*=all
-com.ibm.ws.logging.max.file.size=20
-com.ibm.ws.logging.max.files=10
-com.ibm.ws.logging.trace.format=BASIC
