--- dev/io.openliberty.webcontainer.servlet.6.1.internal_fat/publish/servers/servlet61_RequestCookieHeaderTest/server.xml
@@ -1,20 +1,0 @@
-<!--
-    Copyright (c) 2026 IBM Corporation and others.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License 2.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-2.0/
-    
-    SPDX-License-Identifier: EPL-2.0
- -->
-<server description="Server for testing servlet request COOKIE header: delimiter and surrounding quotes">
-
-    <include location="../fatTestPorts.xml"/>
-
-    <featureManager>
-        <feature>servlet-6.1</feature>
-    </featureManager>
-    
-    <applicationManager autoExpand="true"/>
-
-</server>
